# cmd/daemon/openapi.yaml
openapi: 3.0.0
info:
  title: Gorrent API
  description: Headless torrent search and download daemon for homelab automation.
  version: 1.9.0
servers:
  - url: http://localhost:7800
    description: Gorrent Daemon
security:
  - ApiKeyAuth: []
  - ApiKeyQuery: []
paths:
  /api/search:
    get:
      summary: Search for torrents across enabled scrapers.
      parameters:
        - name: q
          in: query
          description: Search query string
          required: true
          schema:
            type: string
        - name: source
          in: query
          required: false
          description: >
            Restrict search to a single scraper. Valid values: yts, nyaa,
            piratebay, 1337x, eztv, subsplease, fitgirl, torrentscsv, rutracker,
            bittorrented
          schema:
            type: string
      responses:
        '200':
          description: Successful search
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TorrentResult'
        '401':
          description: Unauthorized — API key missing or invalid
      operationId: getApiSearch

  /api/download:
    post:
      summary: Add a torrent to download.
      description: >
        Provide either a 'magnet' link or bare 40-char infohash directly, or an
        'auto' query to automatically search and pick the best result.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                magnet:
                  type: string
                  description: Full magnet URI or bare 40-character infohash
                auto:
                  type: string
                  description: Search query — Gorrent will auto-pick the highest-scored result
                category:
                  type: string
                  description: Category name (e.g. movies, tvshows, anime) — routes files to
                    category_dirs
                source:
                  type: string
                  description: Restrict the auto-search to a specific scraper (e.g. nyaa, yts)
                seed_time:
                  type: string
                  description: Custom seed duration ("30d", "2h", "0" for infinite). Overrides daemon max_seed_days.
                callback:
                  type: string
                  description: >
                    Webhook URL. Gorrent will POST {event, name, hash} to this
                    URL when the download reaches 100%.
      responses:
        '200':
          description: Download started
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: started
                  magnet:
                    type: string
        '400':
          description: Bad request — missing magnet or auto parameter
        '401':
          description: Unauthorized
        '404':
          description: No results found for auto-download query
      operationId: postApiDownload

  /api/seed:
    post:
      summary: Turn a local path into a torrent and seed it.
      description: >
        Takes an existing file or directory path on the local filesystem,
        generates the .torrent metainfo, saves it to download_dir, and begins
        seeding it directly to peers without re-downloading.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - path
              properties:
                path:
                  type: string
                  description: Absolute path to the local file or directory to seed
                category:
                  type: string
                  description: Optional category
                seed_time:
                  type: string
                  description: Custom seed duration ("30d", "2h", "0" for infinite)
      responses:
        '200':
          description: Seeding successfully started
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: seeding
                  info_hash:
                    type: string
                  name:
                    type: string
                  total_bytes:
                    type: integer
                  magnet:
                    type: string
                  torrent_file:
                    type: string
        '400':
          description: Invalid path or missing parameter
        '401':
          description: Unauthorized
      operationId: postApiSeed

  /api/status:
    get:
      summary: Get status of all active downloads.
      responses:
        '200':
          description: List of active torrents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TorrentStatus'
        '401':
          description: Unauthorized
      operationId: getApiStatus

  /api/torrent:
    delete:
      summary: Stop and drop an active torrent (files remain on disk).
      parameters:
        - name: hash
          in: query
          description: 40-character infohash of the torrent to stop
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Torrent stopped successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: stopped
                  hash:
                    type: string
        '401':
          description: Unauthorized
        '404':
          description: Torrent not found
      operationId: deleteApiTorrent

  /api/torrent/seed-time:
    post:
      summary: Update the seed duration limit for an active or completed torrent.
      description: Set custom seed duration (e.g. "30d", "2h", "0" for infinite, or "" to clear override).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - hash
              properties:
                hash:
                  type: string
                  description: 40-character infohash
                seed_time:
                  type: string
                  description: Seed duration ("30d", "2h", "0" for permanent, "" to clear)
      responses:
        '200':
          description: Seed limit successfully updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  hash:
                    type: string
                  seed_time:
                    type: string
        '400':
          description: Invalid seed_time or missing hash
        '401':
          description: Unauthorized
        '404':
          description: Torrent not found
      operationId: postApiTorrentSeedTime

  /api/ws:
    get:
      summary: WebSocket endpoint — streams TorrentStatus array every 1 second.
      responses:
        '101':
          description: Switching Protocols (WebSocket upgrade)
      security:
        - ApiKeyAuth: []
        - ApiKeyQuery: []
      operationId: getApiWs

  /api/docs:
    get:
      summary: Returns this OpenAPI specification as YAML.
      security: []
      responses:
        '200':
          description: OpenAPI YAML document
          content:
            application/yaml:
              schema:
                type: string
      operationId: getApiDocs

  /health:
    get:
      summary: Docker healthcheck endpoint.
      security: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok
      operationId: getHealth

  /metrics:
    get:
      summary: Prometheus metrics endpoint (no auth required).
      security: []
      responses:
        '200':
          description: Prometheus text format
          content:
            text/plain:
              schema:
                type: string
              example: |
                gorrent_torrents_active 2
                gorrent_bytes_downloaded 104857600
                gorrent_bytes_uploaded 52428800
      operationId: getMetrics

  /files/{path}:
    get:
      summary: Stream a downloaded file with HTTP Range support.
      description: >
        Serves files from the configured download_dir over HTTP. Supports the
        Range header for partial content (ideal for video streaming in Plex,
        Jellyfin, or a browser). Requires authentication if api_key is set. Path
        is relative to download_dir (e.g. /files/Movie.Name/video.mkv).
      parameters:
        - name: path
          in: path
          required: true
          description: Relative path to the file under download_dir
          schema:
            type: string
      responses:
        '200':
          description: Full file content
        '206':
          description: Partial content (Range request)
        '401':
          description: Unauthorized
        '404':
          description: File not found
      operationId: getFilesByPath

components:
  schemas:
    TorrentResult:
      type: object
      properties:
        infoHash:
          type: string
        name:
          type: string
        sizeBytes:
          type: integer
        seeders:
          type: integer
        leechers:
          type: integer
        source:
          type: string
          description: Scraper that returned this result
        magnet:
          type: string
        score:
          type: integer
          description: Composite score (seeders * weight) used for auto-picking

    TorrentStatus:
      type: object
      properties:
        hash:
          type: string
          description: 40-character infohash
        name:
          type: string
        downloaded:
          type: integer
          description: Bytes downloaded so far
        length:
          type: integer
          description: Total size in bytes
        peers:
          type: integer
          description: Number of active peer connections
        seed_time:
          type: string
          description: Custom seed limit ("30d", "2h", "0")
        completed_at:
          type: integer
          description: Unix timestamp in seconds when download completed
        seed_remaining_sec:
          type: integer
          description: Remaining seed duration in seconds (-1 for infinite)
        seed_until:
          type: integer
          description: Unix timestamp in seconds when seed limit expires

  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
    ApiKeyQuery:
      type: apiKey
      in: query
      name: apikey